requirements.txt
allure-pytest==2.12.0
pytest==7.1.2
selenium==4.1.5
requests==2.27.1
pytest-xdist==2.5.0